pyproject.toml
[tool.poetry]
name = "student"
version = "0.1.0"
description = ""
authors = ["Priyanshu Tripathi <priyanshu.tripathi@volkswagen.co.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
